requirements.txt
# MCP Server dependencies
mcp[sse]>=1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
starlette>=0.27.0
httpx>=0.25.0
python-dotenv>=1.0.0
load_dotenv>=0.1.0

# Optional: for better async performance
aiofiles>=23.0.0